reorg command line utilities
